Add new CSS variables, mixins, styles and update layout
This commit defines new CSS variables and mixins in setup.scss, variables-derived.scss, and mixins.scss respectively. It also includes their respective usage in generic.scss. In addition to that, a navigation aside was added to the layout master template for better user accessibility and site navigation. These changes aim to provide more styling options to ensure a consistent look and feel across the application and improved user experience.