Add navigation bar height and sidebar width to SCSS variables
In this commit, the height of the navigation bar and the width of the sidebar have been defined in the SCSS variables file. Also, in the navbar.scss file, navbar height value has been referenced from SCSS variables. These changes will help in maintaining a consistent layout across different components.