Updated scrollbars with new color scheme
The scrollbar color scheme was updated with new variables to make it more consistent with the rest of the application. This change was applied across the dark, light and setup themes for consistency. The colors for the scrollbar were fetched using the newly defined 'schollbar', 'schollbar-hover' variables. An extra 'scrollbar-color' attribute was added to the 'aside.scss' layout with the new scheme.