Refactor SCSS theme and layout for sidebar and color variables
The commit adds secondary and tertiary lightness levels to the dark theme design and uses them to adjust the sidebar background property. Color definitions in the variables-init SCSS files are refactored to use scheme hue and saturation for better consistency. Additionally, the sidebar layout has been modified to enhance the brand display, including adjustments to font size and color, and addition of a gap for better spacing. Lastly, the master,tmpl layout is tweaked to include a 'expand-xl' class to the nav tag within the sidebar.