Update dark theme implementation and refine SCSS variables and functions
Reworked the dark theme implementation in the `_master.tmpl` file and updated the SCSS for better handling. Adjusted styles within the `dark.scss` file so that they are applied when the data-theme attribute is set to dark. Also, improved the handling of numbers and shades in the SCSS function and variable files for more efficient and intuitive usage.