Add loading state styles to buttons
The styles for loading state in the button element have been added. Now, when a button is in the loading state, the text color becomes transparent, the button loses its box-shadow, and no pointer events are accepted. A loader is also added right in the center to indicate the loading status.