Update styling for file form inputs
In the scss/forms/file.scss file, several updates were made to improve styling for file form inputs. This includes modifying the flex direction, height, padding, and border-width for certain components as well as adjusting the size of icons. This would make the presentation of file form inputs more consistent and visually engaging.