Add properties and modifiers to form control in SCSS
This change introduces several properties and modifiers to the form control. The CSS properties include position, clear, text-align, box-sizing, font-size, and top. The use of the `{vi.$prefix}` has been expanded for better styling of form elements. Modifiers including `.has-icon-left`, `.has-icon-right`, `.is-small`, `.is-medium`, and `.is-large` have been added for more form control customization.