Update layout header with classes
Modified the CSS styles for the header layout, adding more classes for better structure and functionality. Adjustments have been made to the properties of header-brand and header-tabs, and padding has been added to the body and html when the header is fixed at either the top or bottom. Also introduced new classes for header-burger, header-menu, header-item, header-link, header-content, header-dropdown and header-divider.