Add preview section to README
A Preview section has been added to the README. This section not only provides a brief description of the responsive and user-friendly design of 'Accouter', but also an illustrative image of the same. Details regarding possible customization and the meticulous attention to detail in crafting each component are also included.