Add new features list to README
This commit adds a list of new product features to the README document. The list includes in-depth descriptions of all added features and services, along with their benefits, to provide visitors with a clear understanding of our product. The goal is to enhance user engagement and guide potential users through the features of the admin panel.